Add configuration service and move backend URL to appsettings file
In this commit, we added a configuration service to allow for easier management of configuration values. The backend URL, previously hardcoded in the PostRepository file, is now being pulled from the appsettings.json file. This change will make it easier to modify the backend URL without needing to modify the code itself.

Took 23 minutes